requirements.txt
peewee
python3-discogs-client
jinja2
click
toml
